+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -3,7 +3,7 @@
  */
 import 'jsdom-global/register';
 import jwt from 'jsonwebtoken';
-import {cloneDeep, forEach, isEqual} from 'lodash';
+import {cloneDeep, forEach, isEqual, isUndefined} from 'lodash';
 import sinon from 'sinon';
 import * as internalMediaModule from '@webex/internal-media-core';
 import StateMachine from 'javascript-state-machine';
@@ -411,6 +411,21 @@ describe('plugin-meetings', () => {
           });
         });
       });
+
+      describe('#isLocusCall', () => {
+        it('returns true if it is a call', () => {
+          meeting.type = 'CALL';
+
+          assert.isTrue(meeting.isLocusCall());
+        });
+
+        it('returns false if it is not a call', () => {
+          meeting.type = 'MEETING';
+
+          assert.isFalse(meeting.isLocusCall());
+        });
+      });
+
       describe('#invite', () => {
         it('should have #invite', () => {
           assert.exists(meeting.invite);
@@ -1210,6 +1225,8 @@ describe('plugin-meetings', () => {
             Media.createMediaConnection,
             false,
             meeting.getMediaConnectionDebugId(),
+            webex,
+            meeting.id,
             sinon.match({turnServerInfo: undefined})
           );
           assert.calledOnce(meeting.setMercuryListener);
@@ -1280,6 +1297,8 @@ describe('plugin-meetings', () => {
             Media.createMediaConnection,
             false,
             meeting.getMediaConnectionDebugId(),
+            webex,
+            meeting.id,
             sinon.match({
               turnServerInfo: {
                 url: FAKE_TURN_URL,
@@ -1531,6 +1550,8 @@ describe('plugin-meetings', () => {
             Media.createMediaConnection,
             false,
             meeting.getMediaConnectionDebugId(),
+            webex,
+            meeting.id,
             sinon.match({
               turnServerInfo: {
                 url: FAKE_TURN_URL,
@@ -1750,15 +1771,15 @@ describe('plugin-meetings', () => {
             });
         };
 
-        const checkMediaConnectionCreated = ({mediaConnectionConfig, localTracks, direction, remoteQualityLevel, expectedDebugId}) => {
+        const checkMediaConnectionCreated = ({mediaConnectionConfig, localTracks, direction, remoteQualityLevel, expectedDebugId, meetingId}) => {
           if (isMultistream) {
             const {iceServers} = mediaConnectionConfig;
 
-            assert.calledOnceWithExactly(multistreamRoapMediaConnectionConstructorStub, {
+            assert.calledOnceWithMatch(multistreamRoapMediaConnectionConstructorStub, {
               iceServers,
               enableMainAudio: direction.audio !== 'inactive',
               enableMainVideo: true
-            }, expectedDebugId);
+            }, meetingId);
 
             Object.values(localTracks).forEach((track) => {
               if (track) {
@@ -1800,6 +1821,7 @@ describe('plugin-meetings', () => {
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
+            meetingId: meeting.id
           });
 
           // and SDP offer was sent with the right audioMuted/videoMuted values
@@ -1828,7 +1850,8 @@ describe('plugin-meetings', () => {
               screenShareVideo: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
-            expectedDebugId
+            expectedDebugId,
+            meetingId: meeting.id
           });
 
           // and SDP offer was sent with the right audioMuted/videoMuted values
@@ -1860,6 +1883,7 @@ describe('plugin-meetings', () => {
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
+            meetingId: meeting.id
           });
 
           // and SDP offer was sent with the right audioMuted/videoMuted values
@@ -1888,7 +1912,8 @@ describe('plugin-meetings', () => {
               screenShareVideo: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
-            expectedDebugId
+            expectedDebugId,
+            meetingId: meeting.id
           });
 
           // and SDP offer was sent with the right audioMuted/videoMuted values
@@ -1917,7 +1942,8 @@ describe('plugin-meetings', () => {
               screenShareVideo: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
-            expectedDebugId
+            expectedDebugId,
+            meetingId: meeting.id
           });
 
           // and SDP offer was sent with the right audioMuted/videoMuted values
@@ -5667,6 +5693,60 @@ describe('plugin-meetings', () => {
           waitingForOthersToJoinSpy.restore();
         });
 
+
+        forEach(
+          [
+            {
+              actionName: 'canShareApplication',
+              callType: 'CALL',
+              expectedEnabled: true,
+            },
+            {
+              actionName: 'canShareApplication',
+              callType: 'MEETING',
+              expectedEnabled: false,
+            },
+            {
+              actionName: 'canShareDesktop',
+              callType: 'CALL',
+              expectedEnabled: true,
+            },
+            {
+              actionName: 'canShareDesktop',
+              callType: 'MEETING',
+              expectedEnabled: false,
+            },
+            {
+              actionName: 'canShareContent',
+              callType: 'CALL',
+              expectedEnabled: true,
+            },
+            {
+              actionName: 'canShareContent',
+              callType: 'MEETING',
+              expectedEnabled: false,
+            },
+          ],
+          ({actionName, callType, expectedEnabled}) => {
+            it(`${actionName} is ${expectedEnabled} when the call type is ${callType}`, () => {
+              meeting.type = callType;
+              meeting.setUpLocusInfoMeetingInfoListener();
+
+              const callback = locusInfoOnSpy.thirdCall.args[1];
+
+              const payload = {
+                info: {
+                  userDisplayHints: [],
+                },
+              };
+
+              callback(payload);
+
+              assert.equal(meeting.inMeetingActions.get()[actionName], expectedEnabled);
+            });
+          }
+        );
+
         forEach(
           [
             {
@@ -5704,10 +5784,26 @@ describe('plugin-meetings', () => {
               requiredDisplayHints: [],
               requiredPolicies: [SELF_POLICY.SUPPORT_FILE_TRANSFER],
             },
+            {
+              actionName: 'canShareDesktop',
+              requiredDisplayHints: [DISPLAY_HINTS.SHARE_DESKTOP],
+              requiredPolicies: [],
+              enableUnifiedMeetings: false,
+            },
+            {
+              actionName: 'canShareApplication',
+              requiredDisplayHints: [DISPLAY_HINTS.SHARE_APPLICATION],
+              requiredPolicies: [],
+              enableUnifiedMeetings: false,
+            },
           ],
-          ({actionName, requiredDisplayHints, requiredPolicies}) => {
+          ({actionName, requiredDisplayHints, requiredPolicies, enableUnifiedMeetings}) => {
             it(`${actionName} is enabled when the conditions are met`, () => {
               meeting.selfUserPolicies = {};
+
+              meeting.config.experimental.enableUnifiedMeetings = isUndefined(enableUnifiedMeetings)
+                ? true
+                : enableUnifiedMeetings;
 
               forEach(requiredPolicies, (policy) => {
                 meeting.selfUserPolicies[policy] = true;
@@ -5729,7 +5825,6 @@ describe('plugin-meetings', () => {
             });
 
             if (requiredDisplayHints.length !== 0) {
-              
               it(`${actionName} is disabled when the required display hints are missing`, () => {
                 meeting.selfUserPolicies = {};
 
@@ -5751,9 +5846,7 @@ describe('plugin-meetings', () => {
 
                 assert.isFalse(meeting.inMeetingActions.get()[actionName]);
               });
-
             }
-
 
             it(`${actionName} is disabled when the required policies are missing`, () => {
               meeting.selfUserPolicies = {};